[dynamic-exports] fix incorrect conversion of unsound anys to untyped anys
Summary:
https://github.com/facebook/flow/issues/7473 pointed out that the `dynamic-export` lint was firing more often than it should, particularly in cases like

```
export function a () {}
```
or
```
export class A {}
```

This had to do with the fact that we modeled `static`s  as `Untyped` rather than as their own unsoundness kind, as well as the fact that during `assert_ground` we scrubbed function types to have `Untyped` `any` kinds instead of `Unsound`. This changes that behavior so that these cases are properly ignored during calculation of the lint.

Reviewed By: jbrown215

Differential Revision: D14096453

fbshipit-source-id: 26b610a55ee4a216cad008589a707048aea1cd47