Make React.Ref contra-variant.
Elements only write to refs. This worked perfectly fine with functions refs. But in the case of the useRef style objects, there can be some problems.

By making the `current` key write only the type become contra-variant again. This means that you can pass the ref for A | B to an element that will give you A.

NOTE: this change will also require changes to the implementation of React$AbstractComponent to make the type write-only as well.