[react] Make current write-only in React.Ref
Summary:
Making the ref write-only allows union types for refs. See the test case for an example.

To build an intuition for why this is correct, consider how the code would look for `React.createElement`. After the element is created, React should write the current field of the passed in `ref` with the instance of the component. React does not need to read that ref anywhere!

For more confidence, consider that the ref can also be a function where the argument is write-only by the nature of being a function argument.

Reviewed By: samwgoldman

Differential Revision: D14169874

fbshipit-source-id: 4fb2846c91143b5e515c8e9783f61663cff0c401