[trusted-types] Initial implementation of trust data
Summary: This diff sets up the `trust.ml` module for actual use (as opposed to being a placeholder before). Trust data describes whether a type is trusted, private, both, or neither--see the discussion in `trust.ml` for details. The details of `trust.ml`'s implementation should be fine to keep hidden; instead, trust information can be generated with several new trust constructors which (in later diffs) will be used in places like `statement.ml` to assign trust to types.

Reviewed By: dsainati1

Differential Revision: D13874139

fbshipit-source-id: ee5dfe73a03a22292989eb40a9c8c8cb568bb2f3