[PR] [Docs] Add common mistakes page
Summary:
I observed that newcomers (based on this repo issues and difficulties that people in my team have) tend to encounter the same 4/5 issue. Perhaps a "common mistakes", troubleshooting or something like that could be helpful. So here is a simple proposition :)

I divided this section in two parts, firstly, the problems that people have by not understanding enough Flow's type system, and secondly, flow current limitations.

I struggle to understand how the website is built, so I just added the page, feel free to show me the right direction to make this change work!

Lastly, I saw that there is a "tips" folder in the docs, perhaps this page would better belong there, however, I don't think it's published, so I preferred to add it to the more common `lang` folder.
Pull Request resolved: https://github.com/facebook/flow/pull/4554

Reviewed By: nmote

Differential Revision: D8889955

Pulled By: jbrown215

fbshipit-source-id: adb246727c04c301a850a6733e797a8f5ccbb896