Disallow importing a function as a type
Summary:
We used to allow function names to be used as types, in order to support ES3-style constructors. Now that classes are in widespread use, this functionality was dropped in D9725084. However, the case addressed here appears to have been missed, so this diff makes Flow's behavior more consistent and closes some holes.

Currently, Flow disallows you from using a function name as a type, unless you export the function (as a value) and then import it somewhere else as a type. This closes that hole and makes things consistent. See D14318606 (the diff below), and the changes to the test output here, for a concrete example.

Reviewed By: samwgoldman

Differential Revision: D14318241

fbshipit-source-id: 936cbd2c45bd01f9db506e5eb3fb0a1bfec862a7