Saved state + lazy mode should be lazy
Summary:
Loading from saved state is somewhat lazy already. It only rechecks the files which have changed since the saved state was generated. But thanks to the fanout problem, this could be half the repository.

This diff makes saved state even lazier when using lazy mode. Instead of doing a full check of all the changed files, we only re-parse them and update the dependency graph. This side-steps the fanout problem, since we don't need to look at dependents or dependencies.

Reviewed By: mroch

Differential Revision: D14034701

fbshipit-source-id: e9b8444767c4905e29ba339ca5de9cd26bda1b65