Split recheck in twain
Summary:
Currently, when you start a server using saved state and in lazy mode, it will recheck every file that has changed since the saved state has changed.

I am changing the saved state lazy behavior to only parse & update the dependency graph for the changed files. So I need to be able to call the first half of `recheck` without calling the second half.

The API would be a little messy, since the first half of recheck computes some things which are also useful for the second half. So I encapsulated it into the new `Recheck` module.

Reviewed By: mroch

Differential Revision: D14034702

fbshipit-source-id: 903500d679d07e1bdd8b1ec8e8c5165e0b8b2d6c