delet the garbage collector
Summary: A previous diff moved the location that we perform garbage collection in `merge_component_strict`, but this caused a crash when encountering a cycle. However, we missed this because we only actually run garbage collection in debug mode, as it is a holdover from a previous architecture. This removes the garbage collector entirely.

Reviewed By: mroch

Differential Revision: D14430747

fbshipit-source-id: b6445f2a6a2e8d490eb072cd1a1baf33b563be9b