delet the garbage collector
Summary: A previous diff moved the location that we perform garbage collection in `merge_component_strict`, but this caused a crash when encountering a cycle. However, we missed this because we only actually run garbage collection in debug mode, as it is a holdover from a previous architecture. This removes the garbage collector entirely.

Reviewed By: mroch

Differential Revision: D14430747

fbshipit-source-id: 41f87d15a540230ab22762c055a12336ff08fb7b