[Fix] do not report LintErrors in unchecked files
Summary:
Before, we reported `LintError`s produced during parsing/signature-building of a **non-checked** file
(under a lazy server that is). This was the case for both suppressed and unsuppressed lint errors.

This change ensures that the `LintError`s we report belong to files that are checked. The reason
this is necessary is because we do not include unchecked files in the suppression mapping and
so suppressions in those files are not respected.

Reviewed By: samwgoldman

Differential Revision: D14380862

fbshipit-source-id: ae36bc0357e1a0f29fcf5e170f97bd47a16bf341