improved error messages for importing values as types
Summary: This improves the error messages Flow generates when it sees a value getting imported as a type. First, it makes a slight grammar change to get rid of a sentence fragment. Second, when the error message is generated by importing a default export, this diff changes the wording to avoid the awkward formation "the value the default export".

Reviewed By: nmote

Differential Revision: D14414227

fbshipit-source-id: 5753928ca41ab9145c70c81b72cd34fbd405f928