Add JSON.stringify overload which takes `mixed`
Summary:
https://github.com/facebook/flow/pull/7447 / D13975675 made this overly strict. We thought it would be fine but it is proving to lead to some unnecessary rollout pain.

This change adds a `JSON.stringify` overload that takes `mixed` and returns `string | void`. This leads to more precision than the original PR without sacrificing any soundness.

While testing this change, I found another issue (see the TODO in the newly-added test), but this issue was an issue with the original pre-#7447 behavior, and with the post #7447 behavior. This diff is still strictly an improvement.

Reviewed By: mroch

Differential Revision: D14468687

fbshipit-source-id: 67319127633e0272d11aaeeacb939bbb3e874a74