[hack] Fail fast if fail to load saved-state
Summary:
The big change here is that if serverInit detects that saved-state init had failed, then it now checks whether hh.conf "require_mini_state" flag is true. If it is, then it terminates immediately.

I've also changed the way the TLS bug is reported. If there was a TLS bug, then ServerLazyInit.saved_state_init has already produced a warning message which describes the bug and suggests workarounds. There's no need for serverInit to scrape the error string as well.

NOTE: the design of the monitor is that if you invoke `hh_server .` and then the server process dies, the monitor REMAINS RUNNING even though it notes that the server is in a "Died_unexpectedly" state. So invoking `hh_server .` isn't a good way of getting an exit code upon failed-to-load-saved-state. That will come later in the stack, when we change `hh_client` and `find_hh.sh` to give useful exit codes.

Reviewed By: 2BitSalute

Differential Revision: D13542300

fbshipit-source-id: 53cfbed0c50682ae61b82f1d121ce936f7cf4035