FIX: Skip bogus TypeApp reason when caching
Summary:
`Cache.PolyInstantiation` caches results for type application, using the following keys:
a. the location of the type app as recorded in the application reason,
b. the type-parameter that is instantiated, and
c. some use_op reason

Unfortunately, (a) is not always reliable since there are code paths that produce
TypeAppT constructors without a code-level type application, e.g.
`canonicalize_imported_type` in Flow_js.

In
```
// lib.js
export class C<X> {
  m(f: X => void) {}
}

// main.js
import type {C} from './lib';
(x: C<number> | C<string>) => {
  x.m(y => {
    (y: number); // error string ~> number
  });
}
```
this caused `C<number>` and `C<string>` to hash to the same entry. Specifically, in this
context, `C<number>` is represented as:
```
TypeAppT (r1, PolyT ([P], ClassT (TypeAppT (r2, PolyT (ThisClassT (..)), [P]))), [number])
              \______________________________________________________________/
                      This part added by `canonicalize_imported_type`
```
`r1` correctly captures the unique/cacheable source code application. `r2` is connected with
the non-unique (across applications) import site, and so (after reducing the above) is incorrectly
used as key to `Cache.PolyInstantiation`.

The result of this in the example was that:
* there was an error between the two type annots `C<number>` and `C<string>`
* there was no error in the cast.

This diff disallows reasons of phony `TypeAppT`s to be used for caching.

Reviewed By: avikchaudhuri

Differential Revision: D14453064

fbshipit-source-id: 454797809bc9ebc72c20571c9b88cc4622fc20f8