Fix `flow lsp` log flushing
Summary:
When `lspCommand` is connected to the server but doesn't receive any message from the IDE or server for 1 second, it emits a `Tick` event. Every `Tick` event we call `EventLogger.flush`. Unfortunately, we never emit `Tick` events. This is because the Flow server sends a status update every second.

Instead, let's just flush the logs asynchronously every 5 seconds or so. This is what the monitor does and it works fine there.

Reviewed By: samwgoldman

Differential Revision: D14446951

fbshipit-source-id: 9af36f989ddeabe72c17d2ded382400172717834