[normalizer] React.createClass
Summary:
Before, type-at-pos on `Foo` in
```
const Foo = React.createClass({
  propTypes: {
    foo: React.PropTypes.string,
    bar: React.PropTypes.string.isRequired
  },
  getInitialState() {
    return { baz: 0 };
  },
  getDefaultProps() {
    return {
      foo: "foo"
    };
  }
});
```
returned
```
Class<React$Component>
```

With this change it returns:
```
Class<React$Component<{ bar: string, foo?: string }, {baz: number}>> & {
  +propTypes: {|
    bar: React$PropType$Primitive$Required<string>,
    foo: React$PropType$Primitive<string>
  |},
  defaultProps: { foo: string }
}
```

Implementation-wise, the normalizer:
1. Looks up `"props"` on the newly create instance to fill in the `Props` type
argument to `React$Component`. Will replace with `any` if it's not found.
2. Looks up `"state"` on the newly create instance to fill in the `State` type
argument to `React$Component`. Converts to an inexact version of that type.
Will replace with `any` if it's not found.
3. Looks up and includes all static fields of the instance.

Reviewed By: samwgoldman

Differential Revision: D14486735

fbshipit-source-id: 0ea4259d81b67c1015350878bb41e8badb583fb6