[trusted-types] Add an (unused) options flag for trusted types
Summary: Adds a CLI flag and .flowconfig option to enable trusted types. Right now this is silent and undocumented (hence the `~docs:""` bit in the CLI flag, which disables printing as an option with `flow --help`), and it doesn't actually affect anything since at this point there isn't really anything to affect. The `--trust-mode` flag and `trust_mode` .flowconfig option both take an argument that is one of "check", "silent", and "none"; "none" is the default. The intent is that "none" will do no checking (ignore the trust bits everywhere), "check" will enable trusted types, and "silent" will do all the work required for trusted types without actually ever raising errors--this is suitable for studying the performance impact of trusted types without having to make sure there's an error diff of zero before the results are useful. The current trust mode setting is stored in the context.

Reviewed By: dsainati1

Differential Revision: D13883298

fbshipit-source-id: 3d2586f16fda00a01e70ba4c35bf7a2f66cb1547