[trusted-types] Initial implementation of trust checking
Summary:
Adds the initial core algorithm for trusted types. When the --trust-mode (cli) / trust_mode (.flowconfig) flag is set to "check", the `trust_flow` function in `flow_js.ml` will be invoked on every flow constraint. This function ignores the types of the lower and upper bounds of the constraint, but instead looks at their trust, and raises an error if the trust is incompatible, using the new `trust_subtype` function in `trust.ml`. (Future diffs improve the error messages.)

This change necessitates a couple other interesting changes. The largest one is that `quick_subtype` in `type.ml` now has to be trust-aware, so that trust enforcement is not short-circuited away. However, we only want to change its behavior when trust is enabled, so we pass in a boolean argument to determine whether trust should be checked.

This diff also adds a bunch of test cases, including some that currently do *not* behave as expected: these are currently located in the /tests/trust/_failing/ subdirectory, which is ignored by the .flowconfig.

Reviewed By: dsainati1

Differential Revision: D13874693

fbshipit-source-id: 8e8165a12f5246d391443c287fff720e80e2035d