[trusted-types] Include trust information in debug output
Summary: When debugging trusted types it can be useful to look at whether or not a type is trusted. This diff makes --verbose output include trust infot for DefTs, but only when the context's `trust_mode` flag is set to `CheckTrust` or `SilentTrust` (e.g. no change when it's `NoTrust`)

Reviewed By: dsainati1

Differential Revision: D13891270

fbshipit-source-id: 85655e80213c64d0c4f39e4759f5ed089c3ca9ff