[trusted-types] Improved error reporting for trusted types
Summary:
Improves error reporting for trusted types. This diff adds a new error, `ETrustIncompatibleWithUseOp`, which is used to report trust errors. Currently it reports errors as "X is not a subtype of Y," which I imagine will need improvement, but it's a starting point. When these errors are printed, I intentionally don't do as much manipulation of ordering (e.g. with `flip_contravariant`) as regular type incompatibilities because ordering is now relevant: we really don't want to see "X is not a subtype of any" when the real issue is that any is not a subtype of X.

I've also added new `reason`s to represent trusted and private types, so that types like "trusted number" are reported and we don't get terrible error messages like "number is not a subtype of number."

There is also a fix so that the `Loc` of a type that originates with the annotation `$Trusted<T>` points to the entire annotation, not just the `T`.

Reviewed By: dsainati1

Differential Revision: D14025850

fbshipit-source-id: 4c1f5ef9e79efcd8b04d75a5965e0b146cfc1ffa