[Coverage] disallow running batch-coverage in lazy mode
Summary: Reports from actual users indicate that the results of `batch-coverage` in lazy mode are misleading, as despite reporting the number of files checked due to lazy mode, this does not account for the fact that these files might not be the ones the user is querying for. Ultimately it is not very useful to produce a partial aggregate coverage number, so we will just disallow using `batch-coverage` for servers in lazy mode.

Reviewed By: panagosg7

Differential Revision: D14526373

fbshipit-source-id: 05f44fecbb83c3ea0413f8fd451d0e70a1ea0aea