Pull out get_next_power_of_two util
Summary: I found myself wanting this and realized I had implemented it before. When I searched for it, I found this, and I also found an identical function (with small differences in names and formatting) that I had implemented three years before in the parser. Unfortunately we don't want to introduce dependencies in the parser, so we can't use this shared definition there.

Reviewed By: dsainati1

Differential Revision: D14606638

fbshipit-source-id: 1b8cac70ede51dfd65566f589862b222fa96e726