Fix node_modules filter for lint warnings
Summary:
We don't show lint errors that are positioned in node_modules, since they are
unactionable. The logic was broken in two ways:

1. It didn't work on Windows, because it assumes '/' directory separators
2. It did not respect the node_resolver_dirnames config option

Happily, the existing logic in `Files.is_within_node_modules` handles both
correctly. This diff simply replaces the buggy implementation with the
known-good one. The bulk of this change is threading around the `options`
record.

Reviewed By: gabelevi

Differential Revision: D14158842

fbshipit-source-id: e9695b7b5f43465a8739bf2b873b344b0e46a4a4