ResizableArray
Summary:
I've thought about implementing something like this several times over the years. I was honestly pretty surprised that I couldn't find something like this already.

Specifically, I want to use this to store abstract location tables. The keys are incrementing integers, so an array is perfect, but I want to abstract away the size. I initially used a Map, but it became clear that we needed better performance.

Reviewed By: samwgoldman

Differential Revision: D14606650

fbshipit-source-id: 3a08494a2010975b07ffe001001dd21b0a0e3649