[Coverage][Trusted Types] Add `trust` mode to coverage command
Summary:
This adds a `--trust` option flag to the `flow coverage` command to include information about trusted types in the coverage output. When this flag is enabled, only trusted expressions will be considered covered, while when it is not enabled the command retains its original behavior and considers all checked expressions covered, whether trusted or not.

Trust is computed alongside the coverage kind; the coverage visitor now produces a `Kind.t * Taint.t` pair instead of just a kind. Trust is computed according to the following rules:]

1) All `any`s are untrusted.

2) A `DefT` just uses the associated trust bit to compute trust. Only the taint of the bit is considered; we do not expose any information about public or private trust.

3) Unions and Intersections merge the trust of each of their elements similarly to the process for coverage kinds. `Untainted | Tainted` produces `Tainted`, since as long as at least one branch of a union can be wrong we cannot trust the entire union. Conversely, `Untainted & Tainted` produces `Untainted`, as saying a type is trusted means that we know for sure its type at runtime. A type that has both a trusted type and an untrusted one  can ignore its untrusted type, since we just know that it is false at runtime.

4) For all other cases trust follows the existing behavior of the coverage visitor; entering types as they are seen. The non-`DefT` base cases are simply trusted, since they carry no trust information with them.

Also, while `any` is always assumed to be untrusted, `empty` can be both trusted and untrusted.This allows it to have different behavior with respect to unions and intersections, which is helpful for refinements.

Reviewed By: panagosg7

Differential Revision: D14515652

fbshipit-source-id: 757ccb01d9ed07832f09b94f741918e91085ddd8