[Trusted Types][Coverage] add trust information to batch coverage
Summary: Uses the new trust-sensitive coverage computation to produce additional data for batch-coverage in trust mode. In trust mode, batch-coverage will output **both** the number/percent of trusted expressions and regular covered expressions. The number of covered expressions is just the sum of the number of trusted and untrusted checked expressions.

Reviewed By: panagosg7

Differential Revision: D14525699

fbshipit-source-id: 1da391178235cb10206484202c971e2891b06467