In command line utilities, check that file exists before returning file object to coverage command
Summary: The `flow coverage` command would crash with an OCaml exception when given a file that didn't exist--try `flow coverage sdfsdfasdf`. This diff adds a file existence check to the command utility that returns files to the coverage command (and others) and raises a nicer error if the file does not exist.

Reviewed By: panagosg7

Differential Revision: D14654843

fbshipit-source-id: 3b2c29fd18e996a82b798bead18829d6a2e8dc78