Update Node stream definitions
Summary:
Updates the stream definitions with the latest Node: https://nodejs.org/api/stream.html
There are lots of fixes too around returning `this`, which necessitated updating
other definitions, since they extend some stream class.
A notable change is that `.pipe` is now generic, and returns the stream type
with which it was called.

Reviewed By: gabelevi

Differential Revision: D14683923

fbshipit-source-id: ff589b7e7cd7829cc66343826310b4bdf97cd098