[PR] [DOCS] Add mention about benefits of covariance and pitfalls of destructuring unions
Summary:
Common misconception is that you should be able to pass an array of "stricter" types to an array that accepts wider range of types. This is not true, and oftentimes it causes confusion why it's not possible. It's fairly easy to fix by changing types to be read-only.

Other thing often causing confusion is destructuring object unions.

Feel free to fix potential grammar errors and enhance the explanation!
Pull Request resolved: https://github.com/facebook/flow/pull/7537

Reviewed By: jbrown215

Differential Revision: D14685691

Pulled By: dsainati1

fbshipit-source-id: 878616aa441ff9109f45a9256e8f6502d19c9bed