[Bugbash] mixed refined by array produces $ReadOnlyArray<mixed>
Summary:
It is not safe to refine `mixed` to `Array<mixed>` as this allows for writing arbitrary values to the result of the refinement, and since we only know that the value is an array, we cannot know that it is safe to write a given value to the array. We do know, however, that anything we get out of the array will be `mixed`, so it is safe to type it as a `$ReadOnlyArray<mixed>`.

```
function bad(array: mixed) {
  if (Array.isArray(array)) {
    const problem: Array<mixed> = array;
    problem[1] = 0;
  }
}

bad((["3"] : Array<string>));
```

Fixes #7197 (along with another change Mike is making)

Reviewed By: mvitousek

Differential Revision: D14785979

fbshipit-source-id: eba057af9f874ab3a9619ca8cfcf515b0dec0ee1