Fix additon and logical operators in generics by adding empty flavors
Summary:
This fixes several unsoundnesses in Flow that came from short-circuiting operators in the presence of generics: +, &&, and ||, as well as some other problems with generics and `Empty`.

Basically, to typecheck generic function bodies, Flow tests all the permutations of the generics having either `EmptyT` and their upper bounds. For example, to detect the error here:
```
function f<X, Y:number>(y: Y): X {
  return y;
}
```
Flow tests this with `X=Empty, Y=number`, `X=Mixed, Y=empty`, and `X=Mixed, Y=number`. It fails because in the first case, `number </: Empty`. However, because of the specific logic to some operators, it was possible to write an expression that was always a subtype of the return type. For example,

```
function g<X, Y>(x:X, y: Y): X {
  return x && y;
}
```
In the cases where `X=Mixed`, this is always well-typed. If `X=Empty, Y=Mixed`, the && operation short circuits because Flow models it as `Empty ~> AndT(Mixed, UB)`, and `Empty` can flow into anything via ground subtyping. Furthermore, even if ground subtyping didn't obviate ths operation, `Empty` refines to `Empty` in `type_filter` and as a result only the RHS would flow to the output. As a result it's well typed, even if at runtime it is called as `g(1, "breaks")` which returns "breaks" with type number. Similar issues are possible with addition.

To fix this, we create a special flavor of `EmptyT` which 1. does not get short circuited via ground subtyping and 2. which does not cause addition or logical operations to short-circuit. This flavor, `Zeroed`, is only used in typechecking generics. Since `EmptyT` did not previously have flavors, this necessitates a lot of changes to account for it. Original-flavor `EmptyT` is called `Bottom` and still short-circuits as usual. Because `Zeroed` empty types aren't short circuited by ground subtyping (and in fact need to propagate via e.g. repositioning in order to trigger necessary flows) we need to add a bunch of local rules later on in the `flow_js` match statement that return `()` when `Empty` flows into some specific type.

This also revealed some problems with overloading in `Array.reduce`: previously we didn't error on the new test case `reduce.js` that I added in this diff. Unfortunately though, fixing this reveals another small of errors in www and xplat where the wrong overload of `reduce` is being selected: when inside a bounded generic function, the `reduce` definition of type `(TxT => T) x T => T` is being selected when what is desired is `<U>(UxT => U) x U` where `T` is the type of the generic. This can be overriden by manually passing `U` to the generic call--so `x.reduce(...)` becomes `x.reduce<U>(...)`. I'm not sure this is allowed everywhere, though.*

I had originally thought that instead of introducing flavors to `EmptyT`, I could just add a new flavor to `MixedT`. But we could change the above example so that `X` and `Y` are bounded by `number`, and use different subtypes of `number` at the call site to induce the same error. The only other solution I've come up with is wrapping the upper or lower bounds of generics with a new `t`, but that seems painful :/

Fixes #6671. Fixes #7070.

Reviewed By: avikchaudhuri

Differential Revision: D14716013

fbshipit-source-id: 0c109fdde14b0e13c5727ec76b17869d54e46053