[Bugbash] Mixed refines to read-only object under typeof = object
Summary:
It is not safe to refine `mixed` to `{[string] : mixed}` because this allows writes of any kind into the refined object, and we don't know what type of values the object holds. The only safe operation is to read `mixed` out of the object, so the properties must be read only.

```
function bad (x : mixed) {
  if (typeof x === "object" && x !== null) {
    x.a = 3;
  }
}

let obj : {a : string} = {a : "oops"};
bad(obj); // yikes
```

Reviewed By: mvitousek

Differential Revision: D14799387

fbshipit-source-id: 210ef1d98733b085b202dcbd4e1e83d69e707f08