Add test for #7322
Summary: #7322 is due to an issue with union typechecking. When typechecking a value against a union type, all tvars in the union type are assumed to have been resolved. If there are unresolved tvars, they are `any`'d out. That's why the return types of these methods are considered `any`, and the typechecker issues no errors here.

Reviewed By: samwgoldman

Differential Revision: D14784521

fbshipit-source-id: 664a2ef436a0c98af7466d01e36aace787b853a9