Fix to specific/general assignments
Summary:
Fixes 7618.

Was getting weird unsoundnesses when an object with a type similar to `{x: number | string}` had assignments inside and outside of if statements. See bug report https://github.com/facebook/flow/issues/7618 for details.

Turned out, this is because object assignment was setting *both* the specific and general types of the LHS to be the RHS, clobbering the original general type in the scope. This meant that within a nested scope like that of an if, the type of the attribute is treated as just the RHS of the exterior assignment. This fixes that by making object assignment just update the specific type, not the general type (or rather, sets the general type to a tvar representing the original type of the attribute).

Reviewed By: samwgoldman

Differential Revision: D14802729

fbshipit-source-id: 9c1b7e670728ee2521e3dfc59e829ac6cc60f62e