[React][bugbash] Special case AnnotT in eval_destructor to punt the destructor work to the type inside the annotation
Summary:
Previously, an `AnnotT -> TypeDestructorTriggerT` flow would perform the type destruction before unwrapping the `AnnotT`, and thus would not hit the special cases for `UnionT` and `MaybeT` designed to prevent splitting the union into separate lower bounds. Instead, we should perform the same trick of pushing the destructor down into the contents of the `AnnotT` to ensure that these cases are hit before the type destructor is resolved.

As an example, when dealing with `React.ElementType`, we would generate a `AnnotT(OpenT(UnionT(StrT | AbstractComponentT))) -> TypeDestructorT(GetProps)` flow, with the intention of creating a `ReactKitT(GetProps)` whose bounds were the union of the `StrT` and the `AbstractComponentT`, these being the definition of `React.ElementType` in our libdefs.

However, with the previous behavior, we would resolve the `TypeDestructorT(GetProps)` to a `ReactKitT(GetProps)` first, and then peel off layers of the lower bound type until we had two separate flows from `StrT` and `AbstractComponentT` to the `ReactKitT`. This in effect checked that **both** lower bounds were valid when in reality only one was necessary.

Now, we will peel off layers of the lower bound first until we result in a `UnionT -> TypeDestructorT` flow, which uses the special case to perform the correct behavior.

Reviewed By: samwgoldman

Differential Revision: D14710815

fbshipit-source-id: ec355eb2d8f422298ef162b175841779deaa1aa4