Intern comments in return statements.
Summary:
Intern comments in return statements; the implementation follows quite closely
the plan laid out in the related task.  Two changes are noteworthy:

1. The suggest tool was changed to avoid printing interned comments twice.
This issue was present before my diff but it was not catched by any tests up
to now.

2. A new simple heuristic in `Peek.leading_comments` and
`Peek.trailing_comments` is used to find comments which should likely
be attached to a statement starting or ending at the current location.

For example, the heuristic avoids attaching the comment below as
leading comment of the return statement.

    x = doSomething(y);  // Does something
    return x + 1;

Reviewed By: pakoito

Differential Revision: D14643177

fbshipit-source-id: 52747fa883b4fe206ebf5d3e094861fa7342e141