[normalizer] Option to not evaluate type-destructors
Summary:
This adds an option to the normalizer to avoid evaluating type-destructor
`EvalT`s.

Setting this option to to `true` will cause the normalizer to try to reuse the
cached results of evaluated type-destructors. This often leads to larger types.

If this is set to `false`, then instead the normalizer will try to use:
- a potentially attendant type-alias annotation, or
- reuse the utility type that corresponds to this the specific type-destructor.

Choosing `false` will typically result in smaller produced types, which makes
it a more appropriate option for codemods.

Reviewed By: samwgoldman

Differential Revision: D14238438

fbshipit-source-id: 0993b5da713f6103004075b5e6936228265aa11b