[Ty] Make simplification sensitive to kinds of Bot (empty)
Summary:
Adds an `simplify_empty` option to the normalizer, which determined if it
will consider all kinds of Ty.Bot the same when simplifying types.

The normalized type Ty.Bot may correspond to either the `Empty` type, no
lower-bounds or the internal types `MatchingPropT` or `TypeDestructorTriggerT`.

These types are not easy to normalize, but may still encode some constraint.
When using normalized types for codemods we might want to know if there might
be some constraints that we missing in the normalized type. Setting `simplify_empty`
to false will prevent this important information from being lost in simplifications
of the form:
```
empty | T ~> T
```

To preserve current behavior, this option is set to `true` for type services.

Reviewed By: samwgoldman

Differential Revision: D14763241

fbshipit-source-id: 86d0feafd9832cc275411c60154cb82e0df02a38