Batch DidOpen
Summary:
The Flow server receives DidOpen notifications from the LSP, containing 1 or more files. There are certain situations where, instead of a single notification with N files, we get N notifications with 1 files. This leads to N rechecks.

We should keep on the side the pending files for which we've received DidOpen notifications but haven't processed them yet. Then, the next time we handle a DidOpen notification, we'll include all the pending files in the recheck.

Reviewed By: nmote

Differential Revision: D14875654

fbshipit-source-id: 2bb4db156cb65f93523d3b37d7fd5e31483d3cff