[PR] [DOCS] Update creation.md
Summary:
I've replaced the alert TODO with a clear indication that the documentation is simply not ready yet. Defining modules that are accessed via a relative require/import path is not a well covered area and to find some mention of it in the docs but be presented with an alert box isn't great.

<!--
  If this is a change to library defintions, please include links to relevant documentation.
  If this is a documentation change, please prefix the title with [DOCS].

  If this is neither, ensure you opened a discussion issue and link it in the PR description.
-->
Pull Request resolved: https://github.com/facebook/flow/pull/7544

Reviewed By: nmote

Differential Revision: D14882579

Pulled By: gkz

fbshipit-source-id: 6530a5cab10f02bd2c3221056bf90eb3801b2a30