[normalizer] Rework import information in Ty.t
Summary:
Each symbol in Ty.t types contains information about where it came from.
If it came from another file, it's useful to know if it has been imported to this
file and if so how (as type or as a value).

The normalizer computed this information, by pre-compiling a map of imported
names to their defining (remote) locations. This map was of the form
```
string -> Ty.symbol
```
There are a couple things I didn't like about this:
1. To determine if a symbol encountered during normalization is imported we had
to check for its name in this map and then also check that the def-loc coincided.
The first check was redundant. Def-locs should be enough for this.
2. It didn't really say anything about how the symbol was imported.

This diff changes this by:

1. Reversing the map so that now we map def-locs to import information.
2. We include the mode of import (as value, type, or typeof).

Reviewed By: dsainati1

Differential Revision: D14816951

fbshipit-source-id: d1babd0da9ff7f35b58903a39e060bcd44ac6f35