Smarter version mismatch behaviors
Summary:
I have built 4 possible behaviors for when the client and server have different build IDs.

1. Stop server. The server dies and the client can decide whether to start a new server or not. This is the current behavior
2. Error client. Reject the connection and the client errors.
3. Restart client. Reject the connection. The client will `exec` itself using the same binary as the server and try again.
4. Choose newest. Compare the semver for the client and server. If the client is newer, then stop the server. If the server is newer or if their semvers are the same, restart the client.

At the CLI, the user can choose the behavior using `--on-mismatch`. The default is `--on-mismatch choose-newest`. This way, when we launch a patch release (which does NOT modify the `.flowconfig`), a newer client will kill the older server.

The JSON-based handshake is still backwards compatible with older binaries. When a newer client or server communicates with an older client or server, we maintain the old behavior.

* An old client communicating with a new server will always cause the server to die
* A new client communicating with an old server will always expect the server to die

Reviewed By: ljw1004

Differential Revision: D14644243

fbshipit-source-id: 288d9a8af39711ea50aebf28d106b0a32ea2ae8d