[Lint] ignore deprecated-utility and dynamic-export when applying all=setting rules
Summary: Discussion in issue #7473 indicated that users wanted the `dynamic-export` lint to not be included in the `all=error` setting, so it should now be ignored by the all setting. This also adds `deprecated-utility` to the ignore list, as we don't want users to accidentally enable the use of unsupported types if they choose to turn off other, less severe, lints.

Reviewed By: nmote

Differential Revision: D15002081

fbshipit-source-id: 6d779ec93180804adfde524194fe8a4dcfdaadbf