Refactor object type property fold
Summary:
The code that processes object type properties is pretty complex, mostly due to
spread elements. The fold accumulates properties into a "slice" until we find a
spread element. Before this change, we would convert the slice into a type
eagerly during the fold. After this change, we keep the accumulated data as
either a slice or a spread, only converting to types after the fold.

I also packaged up the `o, os` part of the fold accumulator into a module with a
data type, which makes the fold a bit clearer.

This refactor makes a follow-up fix a bit clearer.

Reviewed By: panagosg7

Differential Revision: D14672824

fbshipit-source-id: b813e6d7a5dfd2b326d3731c04dc4fd4c51e9926