Move object annotation processing code
Summary:
This is just a small refactoring which moves this code outside of the core
pattern in the convert function. I made a few small changes like hoisting
functions into the static scope, which required adding a few parameters which
were previously simply closed over.

This will make it a bit easier to make some follow-up changes.

Reviewed By: panagosg7

Differential Revision: D14669515

fbshipit-source-id: d069113640e0ffbb749ec024e58fd79191ce42de