Remove deprecated $call property syntax
Summary:
This feature was deprecated in D8042447 (Flow v0.75.0), which seems like a more
than reasonable deprecation period. It should be easy to change any uses of this
to use callable property syntax or the [[call]] internal slot property syntax.

Removing this feature makes it possible to clean up a lot of crufty code with
weird corner cases.

Reviewed By: panagosg7

Differential Revision: D14648545

fbshipit-source-id: 48d6aefda5e0952ae6611e7e49f451eef61e89a5