Concretize callable object params before intersection processing
Summary:
When we see an intersection lower bound with a function type upper bound, we
concretize the params to give the speculation machinery a fighting chance to
choose one branch.

We did not do the same thing for callable object upper bounds. This diff adds
more logic to handle that case.

Reviewed By: panagosg7

Differential Revision: D14784956

fbshipit-source-id: 1b9f0da01b8b3c1a7a4cdd532d8a5a1b293621af