Clarify children for fragments and similar React APIs
Summary:
Before this change, a `Fragment`'s children were a non-optional maybe type,
which allows `void` but does not allow omitting the children property entirely.
Instead, the children should be optional. Since `React$Node` already includes
`null`, and `void` is not actually a valid child type, we can entirely replace
the maybe type with optionality.

The fragment expression `<></>` should be an error with a non-optional children
type, since at runtime the element will not be passed a children prop at all.
This worked by accident, since `Fragment` is defined as a function type and
function type's have an `any`-typed statics object, and thus `any` typed default
props.

In a stacked diff, I change function types to use a stricter type for statics
(an empty object with Function.prototype as the __proto__), which exposed this
bug.

Note that this bug also allowed creating a fragment with any children type,
e.g., `<>{() => {}}</>`, even though functions are not members of `React.Node`.

Reviewed By: bvaughn

Differential Revision: D14447780

fbshipit-source-id: db0cf4700cd97762b5c92b8a24ff5e0221c7ff96